Crate subspace_core_primitives

Core primitives for Subspace Network.

Modules

crypto

Various cryptographic utilities used across Subspace Network.

objects

Data structures related to objects (useful data) stored on Subspace Network.

sector_codec

PoR codec that uses polynomial-based systematic erasure coding

Structs

ChunkSignature

VRF signature output and proof as produced by schnorrkel crate.

FlatPieces

Flat representation of multiple pieces concatenated for higher efficient for processing.

LastArchivedBlock

Last archived block

Piece

A piece of archival history in Subspace Network.

PieceIndexHash

Hash of PieceIndex

PublicKey

A Ristretto Schnorr public key as bytes produced by schnorrkel crate.

RewardSignature

A Ristretto Schnorr signature as bytes produced by schnorrkel crate.

Scalar

Representation of a single BLS12-381 scalar value.

SectorId

Data structure representing sector ID in farmer’s plot

Solution

Farmer solution for slot challenge.

U256

256-bit unsigned integer

Enums

ArchivedBlockProgress

Progress of an archived block.

RootBlock

Root block for a specific segment.

Constants

BLAKE2B_256_HASH_SIZE

Size of BLAKE2b-256 hash output (in bytes).

PIECES_IN_SECTOR

How many pieces we have in a sector

PIECES_IN_SEGMENT

128 data records and 128 parity records (as a result of erasure coding).

PIECE_SIZE

Byte size of a piece in Subspace Network, ~32KiB (a bit less due to requirement of being a multiple of 2 bytes for erasure coding as well as multiple of 31 bytes in order to fit into BLS12-381 scalar safely).

PLOT_SECTOR_SIZE

Size of one plotted sector.

PUBLIC_KEY_LENGTH

Length of public key in bytes.

RANDOMNESS_CONTEXT

Randomness context

RANDOMNESS_LENGTH

Byte length of a randomness type.

RECORDED_HISTORY_SEGMENT_SIZE

Recorded History Segment Size includes half of the records (just data records) that will later be erasure coded and together with corresponding witnesses will result in PIECES_IN_SEGMENT pieces of archival history.

RECORD_SIZE

Size of a segment record given the global piece size (in bytes).

REWARD_SIGNATURE_LENGTH

Length of signature in bytes

WITNESS_SIZE

Size of witness for a segment record (in bytes).

Functions

bidirectional_distance

Bidirectional distance metric implemented on top of subtraction

Type Definitions

Blake2b256Hash

BLAKE2b-256 hash output

BlockNumber

Block number in Subspace network.

BlockWeight

BlockWeight type for fork choice rules.

PieceIndex

Piece index in consensus

Randomness

Type of randomness.

RecordsRoot

Records root type.

SectorIndex

Sector index in consensus

SegmentIndex

Segment index type.

SlotNumber

Slot number in Subspace network.

SolutionRange

Type of solution range.