Module subspace_core_primitives::crypto::kzg

Tools for KZG commitment scheme

Structs

Commitment

Commitment to polynomial

Kzg

Wrapper data structure for working with KZG commitment scheme

Polynomial

Commitment to polynomial

Witness

Witness for polynomial evaluation

Constants

EMBEDDED_KZG_SETTINGS_BYTES

Embedded KZG settings as bytes, too big for no_std in most cases Generated with with following command (using current Ethereum KZG Summoning Ceremony):

NUM_G1_POWERS

Number of G1 powers stored in EMBEDDED_KZG_SETTINGS_BYTES

NUM_G2_POWERS

Number of G2 powers stored in EMBEDDED_KZG_SETTINGS_BYTES

Functions

bytes_to_kzg_settings

Function turns bytes into FsKZGSettings, it is up to the user to ensure that bytes make sense, otherwise result can be very wrong (but will not panic).

embedded_kzg_settings

Embedded KZG settings