Function subspace_core_primitives::crypto::kzg::bytes_to_kzg_settings

pub fn bytes_to_kzg_settings(
    bytes: &[u8],
    num_g1_powers: usize,
    num_g2_powers: usize
) -> Result<FsKZGSettings, String>

Function turns bytes into FsKZGSettings, it is up to the user to ensure that bytes make sense, otherwise result can be very wrong (but will not panic).