Module subspace_core_primitives::objects

Data structures related to objects (useful data) stored on Subspace Network.

Mappings provided are of 3 kinds:

for objects within a block

for objects within a piece

for global objects in the global history of the blockchain

Structs

BlockObjectMapping

Mapping of objects stored inside of the block

PieceObjectMapping

Mapping of objects stored inside of the piece

Enums

BlockObject

Object stored inside of the block

GlobalObject

Object stored inside in the history of the blockchain

PieceObject

Object stored inside of the block