subspace_core_primitives

Module pieces

Pieces-related data structures.

Structs

FlatPieces

Flat representation of multiple pieces concatenated for more efficient for processing

Piece

A piece of archival history in Subspace Network.

PieceArray

A piece of archival history in Subspace Network.

PieceIndex

Piece index in consensus

PieceOffset

Piece offset in sector

RawRecord

Raw record contained within recorded history segment before archiving is applied.

Record

Record contained within a piece.

RecordCommitment

Record commitment contained within a piece.

RecordWitness

Record witness contained within a piece.