subspace_core_primitives::pieces

Struct Piece

pub struct Piece(/* private fields */);

A piece of archival history in Subspace Network.

This version is allocated on the heap, for stack-allocated piece see PieceArray.

Internally piece contains a record and corresponding witness that together with segment commitment of the segment this piece belongs to can be used to verify that a piece belongs to the actual archival history of the blockchain.

Implementations

impl Piece

pub const SIZE: usize = 1_048_672usize

Size of a piece (in bytes).

pub fn to_shared(self) -> Self

Ensure piece contains cheaply cloneable shared data.

Internally piece uses CoW mechanism and can store either mutable owned data or data that is cheap to clone, calling this method will ensure further clones will not result in additional memory allocations.