subspace_core_primitives::pieces

Struct PieceArray

pub struct PieceArray(/* private fields */);

A piece of archival history in Subspace Network.

This version is allocated on the stack, for heap-allocated piece see Piece.

Internally piece contains a record and corresponding witness that together with segment commitment of the segment this piece belongs to can be used to verify that a piece belongs to the actual archival history of the blockchain.

Implementations

impl PieceArray

pub fn new_boxed() -> Box<Self>

Create boxed value without hitting stack overflow

pub fn split(&self) -> (&Record, &RecordCommitment, &RecordWitness)

Split piece into underlying components.

pub fn split_mut( &mut self, ) -> (&mut Record, &mut RecordCommitment, &mut RecordWitness)

Split piece into underlying mutable components.

pub fn record(&self) -> &Record

Record contained within a piece.

pub fn record_mut(&mut self) -> &mut Record

Mutable record contained within a piece.

pub fn commitment(&self) -> &RecordCommitment

Commitment contained within a piece.

pub fn commitment_mut(&mut self) -> &mut RecordCommitment

Mutable commitment contained within a piece.

pub fn witness(&self) -> &RecordWitness

Witness contained within a piece.

pub fn witness_mut(&mut self) -> &mut RecordWitness

Mutable witness contained within a piece.

pub fn slice_to_repr(value: &[Self]) -> &[[u8; 1048672]]

Convenient conversion from slice of piece array to underlying representation for efficiency purposes.

pub fn slice_from_repr(value: &[[u8; 1048672]]) -> &[Self]

Convenient conversion from slice of underlying representation to piece array for efficiency purposes.

pub fn slice_mut_to_repr(value: &mut [Self]) -> &mut [[u8; 1048672]]

Convenient conversion from mutable slice of piece array to underlying representation for efficiency purposes.

pub fn slice_mut_from_repr(value: &mut [[u8; 1048672]]) -> &mut [Self]

Convenient conversion from mutable slice of underlying representation to piece array for efficiency purposes.