subspace_core_primitives::pieces

Struct RawRecord

pub struct RawRecord(/* private fields */);

Raw record contained within recorded history segment before archiving is applied.

NOTE: This is a stack-allocated data structure and can cause stack overflow!

Implementations

impl RawRecord

pub const NUM_CHUNKS: usize = 32_768usize

Number of chunks (scalars) within one raw record.

pub const SIZE: usize = 1_015_808usize

Size of raw record in bytes, is guaranteed to be a multiple of ScalarBytes::SAFE_BYTES.

pub fn new_boxed() -> Box<Self>

Create boxed value without hitting stack overflow

pub fn slice_to_repr(value: &[Self]) -> &[[[u8; 31]; 32768]]

Convenient conversion from slice of record to underlying representation for efficiency purposes.

pub fn slice_from_repr(value: &[[[u8; 31]; 32768]]) -> &[Self]

Convenient conversion from slice of underlying representation to record for efficiency purposes.

pub fn slice_mut_to_repr(value: &mut [Self]) -> &mut [[[u8; 31]; 32768]]

Convenient conversion from mutable slice of record to underlying representation for efficiency purposes.

pub fn slice_mut_from_repr(value: &mut [[[u8; 31]; 32768]]) -> &mut [Self]

Convenient conversion from mutable slice of underlying representation to record for efficiency purposes.