subspace_core_primitives::pot
Struct PotOutput

pub struct PotOutput(/* private fields */);

Proof of time output, can be intermediate checkpoint or final slot output

Implementations

impl PotOutput

pub const SIZE: usize = 16usize

Size of proof of time proof in bytes

pub fn derive_global_randomness(&self) -> Randomness

Derives the global randomness from the output

pub fn seed(&self) -> PotSeed

Derive seed from proof of time in case entropy injection is not needed

pub fn seed_with_entropy(&self, entropy: &Blake3Hash) -> PotSeed

Derive seed from proof of time with entropy injection