subspace_core_primitives::solutions

Struct Solution

pub struct Solution<RewardAddress> {
    pub public_key: PublicKey,
    pub reward_address: RewardAddress,
    pub sector_index: SectorIndex,
    pub history_size: HistorySize,
    pub piece_offset: PieceOffset,
    pub record_commitment: RecordCommitment,
    pub record_witness: RecordWitness,
    pub chunk: ScalarBytes,
    pub chunk_witness: ChunkWitness,
    pub proof_of_space: PosProof,
}

Farmer solution for slot challenge.

Fields
public_key: PublicKey

Public key of the farmer that created the solution

reward_address: RewardAddress

Address for receiving block reward

sector_index: SectorIndex

Index of the sector where solution was found

history_size: HistorySize

Size of the blockchain history at time of sector creation

piece_offset: PieceOffset

Pieces offset within sector

record_commitment: RecordCommitment

Record commitment that can use used to verify that piece was included in blockchain history

record_witness: RecordWitness

Witness for above record commitment

chunk: ScalarBytes

Chunk at above offset

chunk_witness: ChunkWitness

Witness for above chunk

proof_of_space: PosProof

Proof of space for piece offset

Implementations

impl<RewardAddressA> Solution<RewardAddressA>

pub fn into_reward_address_format<T, RewardAddressB>( self, ) -> Solution<RewardAddressB>
where RewardAddressA: Into<T>, T: Into<RewardAddressB>,

Transform solution with one reward address type into solution with another compatible reward address type.

impl<RewardAddress> Solution<RewardAddress>

pub fn genesis_solution( public_key: PublicKey, reward_address: RewardAddress, ) -> Self

Dummy solution for the genesis block