Struct subspace_core_primitives::RawRecord

pub struct RawRecord(/* private fields */);

Raw record contained within recorded history segment before archiving is applied.

NOTE: This is a stack-allocated data structure and can cause stack overflow!

Implementations

impl RawRecord

pub const NUM_CHUNKS: usize = 32_768usize

Number of chunks (scalars) within one raw record.

pub const SIZE: usize = 1_015_808usize

Size of raw record in bytes, is guaranteed to be a multiple of Scalar::SAFE_BYTES.

pub fn new_boxed() -> Box<Self>

Create boxed value without hitting stack overflow