Struct subspace_core_primitives::Record

#[repr(transparent)]
pub struct Record(_);

Record contained within a piece.

NOTE: This is a stack-allocated data structure and can cause stack overflow!

Implementations

impl Record

pub const SIZE: usize = 31_648usize

Size of a segment record given the global piece size (in bytes) after erasure coding RawRecord, is guaranteed to be a multiple of Scalar::FULL_BYTES.

pub fn new_boxed() -> Box<Self>

Create boxed value without hitting stack overflow

pub fn slice_to_repr(value: &[Self]) -> &[[u8; 31648]]

Convenient conversion from slice of record to underlying representation for efficiency purposes.

pub fn slice_from_repr(value: &[[u8; 31648]]) -> &[Self]

Convenient conversion from slice of underlying representation to record for efficiency purposes.

pub fn slice_mut_to_repr(value: &mut [Self]) -> &mut [[u8; 31648]]

Convenient conversion from mutable slice of record to underlying representation for efficiency purposes.

pub fn slice_mut_from_repr(value: &mut [[u8; 31648]]) -> &mut [Self]

Convenient conversion from mutable slice of underlying representation to record for efficiency purposes.

pub fn full_scalar_arrays( &self ) -> impl ExactSizeIterator<Item = &[u8; 32]> + '_

Get a stream of arrays, each containing scalar bytes.

pub fn full_scalar_arrays_mut( &mut self ) -> impl ExactSizeIterator<Item = &mut [u8; 32]> + '_

Get a stream of mutable arrays, each containing scalar bytes.