Struct subspace_core_primitives::Solution

pub struct Solution<PublicKey, RewardAddress> {
    pub public_key: PublicKey,
    pub reward_address: RewardAddress,
    pub sector_index: SectorIndex,
    pub total_pieces: NonZeroU64,
    pub piece_offset: PieceIndex,
    pub record_commitment_hash: Scalar,
    pub piece_witness: Witness,
    pub chunk_offset: u32,
    pub chunk: ScalarLegacy,
    pub chunk_signature: ChunkSignature,
}

Farmer solution for slot challenge.

Fields
public_key: PublicKey

Public key of the farmer that created the solution

reward_address: RewardAddress

Address for receiving block reward

sector_index: SectorIndex

Index of the sector where solution was found

total_pieces: NonZeroU64

Number of pieces in archived history at time of sector creation

piece_offset: PieceIndex

Pieces offset within sector

record_commitment_hash: Scalar

Piece commitment that can use used to verify that piece was included in blockchain history

piece_witness: Witness

Witness for above piece commitment

chunk_offset: u32

Chunk offset within a piece

chunk: ScalarLegacy

Chunk at above offset

chunk_signature: ChunkSignature

VRF signature of expanded version of the above chunk

Implementations

impl<PublicKey, RewardAddressA> Solution<PublicKey, RewardAddressA>

pub fn into_reward_address_format<T, RewardAddressB>( self ) -> Solution<PublicKey, RewardAddressB>where RewardAddressA: Into<T>, T: Into<RewardAddressB>,

Transform solution with one reward address type into solution with another compatible reward address type.

impl<PublicKey, RewardAddress> Solution<PublicKey, RewardAddress>where PublicKey: Clone, RewardAddress: Clone,

pub fn genesis_solution( public_key: PublicKey, reward_address: RewardAddress ) -> Self

Dummy solution for the genesis block