Struct subspace_core_primitives::U256

pub struct U256(_);

256-bit unsigned integer

Implementations

impl U256

pub const fn zero() -> Self

Zero (additive identity) of this type.

pub fn one() -> Self

One (multiplicative identity) of this type.

pub fn from_be_bytes(bytes: [u8; 32]) -> Self

Create from big endian bytes

pub fn to_be_bytes(self) -> [u8; 32]

Convert to big endian bytes

pub fn from_le_bytes(bytes: [u8; 32]) -> Self

Create from little endian bytes

pub fn to_le_bytes(self) -> [u8; 32]

Convert to little endian bytes

pub fn checked_add(&self, v: &Self) -> Option<Self>

Adds two numbers, checking for overflow. If overflow happens, None is returned.

pub fn checked_sub(&self, v: &Self) -> Option<Self>

Subtracts two numbers, checking for underflow. If underflow happens, None is returned.

pub fn checked_mul(&self, v: &Self) -> Option<Self>

Multiplies two numbers, checking for underflow or overflow. If underflow or overflow happens, None is returned.

pub fn checked_div(&self, v: &Self) -> Option<Self>

Divides two numbers, checking for underflow, overflow and division by zero. If any of that happens, None is returned.

pub fn saturating_add(&self, v: &Self) -> Self

Saturating addition. Computes self + other, saturating at the relevant high or low boundary of the type.

pub fn saturating_sub(&self, v: &Self) -> Self

Saturating subtraction. Computes self - other, saturating at the relevant high or low boundary of the type.

pub fn saturating_mul(&self, v: &Self) -> Self

Saturating multiplication. Computes self * other, saturating at the relevant high or low boundary of the type.

pub const MIDDLE: Self = _

The middle of the piece distance field. The analogue of 0b1000_0000 for u8.

pub const MAX: Self = _

Maximum value.