Crate subspace_core_primitives

Core primitives for Subspace Network.

Modules

checksum

Module containing wrapper for SCALE encoding/decoding with checksum

hashes

Hashes-related data structures and functions.

objects

Data structures related to objects (useful data) stored on Subspace Network.

pieces

Pieces-related data structures.

pos

Proof of space-related data structures.

pot

Proof of time-related data structures.

sectors

Sectors-related data structures.

segments

Segments-related data structures.

solutions

Solutions-related data structures and functions.

Structs

PublicKey

A Ristretto Schnorr public key as bytes produced by schnorrkel crate.

Randomness

Type of randomness.

ScalarBytes

Single BLS12-381 scalar with big-endian representation, not guaranteed to be valid

U256

256-bit unsigned integer

Constants

REWARD_SIGNING_CONTEXT

Signing context used for creating reward signatures by farmers.

Type Aliases

BlockForkWeight

BlockForkWeight type for fork choice rules.

BlockHash

Block hash in Subspace network.

BlockNumber

Block number in Subspace network.

SlotNumber

Slot number in Subspace network.