Crate subspace_core_primitives

Core primitives for Subspace Network.

Modules

crypto

Various cryptographic utilities used across Subspace Network.

objects

Data structures related to objects (useful data) stored on Subspace Network.

Structs

FlatPieces

Flat representation of multiple pieces concatenated for higher efficient for processing.

LastArchivedBlock

Last archived block

LocalChallenge

A Ristretto Schnorr signature as bytes produced by schnorrkel crate.

Piece

A piece of archival history in Subspace Network.

PieceIndexHash

Hash of PieceIndex

PublicKey

A Ristretto Schnorr public key as bytes produced by schnorrkel crate.

Signature

A Ristretto Schnorr signature as bytes produced by schnorrkel crate.

Solution

Farmer solution for slot challenge.

Enums

ArchivedBlockProgress

Progress of an archived block.

RootBlock

Root block for a specific segment.

Constants

PIECE_SIZE

Byte size of a piece in Subspace Network, 4KiB.

RANDOMNESS_LENGTH

Byte length of a randomness type.

SALT_SIZE

Size of Tag in bytes.

SHA256_HASH_SIZE

Size of Sha2-256 hash output (in bytes)

TAG_SIZE

Size of Tag in bytes.

Functions

bidirectional_distance

Bidirectional distance metric implemented on top of subtraction

Type Definitions

BlockNumber

Block number in Subspace network.

PieceIndex

Piece index in consensus

Randomness

Type of randomness.

Salt

Salt used for creating commitment tags for pieces.

Sha256Hash

Sha2-256 hash output

SlotNumber

Slot number in Subspace network.

Tag

Type of the commitment for a particular piece.