Crate subspace_core_primitives

Core primitives for Subspace Network.

Modules

checksum

Module containing wrapper for SCALE encoding/decoding with checksum

crypto

Various cryptographic utilities used across Subspace Network.

objects

Data structures related to objects (useful data) stored on Subspace Network.

Structs

ArchivedHistorySegment

Archived history segment after archiving is applied.

ChunkWitness

Witness for chunk contained within a record.

FlatPieces

Flat representation of multiple pieces concatenated for more efficient for processing

HistorySize

Size of blockchain history in segments.

LastArchivedBlock

Last archived block

NonEmptyVec

A Vec<> that enforces the invariant that it cannot be empty.

Piece

A piece of archival history in Subspace Network.

PieceArray

A piece of archival history in Subspace Network.

PieceIndex

Piece index in consensus

PieceOffset

Piece offset in sector

PosProof

Proof of space proof bytes.

PosSeed

Proof of space seed.

PotCheckpoints

Proof of time checkpoints, result of proving

PotKey

Proof of time key(input to the encryption).

PotOutput

Proof of time output, can be intermediate checkpoint or final slot output

PotSeed

Proof of time seed

PublicKey

A Ristretto Schnorr public key as bytes produced by schnorrkel crate.

Randomness

Type of randomness.

RawRecord

Raw record contained within recorded history segment before archiving is applied.

Record

Record contained within a piece.

RecordCommitment

Record commitment contained within a piece.

RecordWitness

Record witness contained within a piece.

RecordedHistorySegment

Recorded history segment before archiving is applied.

RewardSignature

A Ristretto Schnorr signature as bytes produced by schnorrkel crate.

SBucket

S-bucket used in consensus

SectorId

Data structure representing sector ID in farmer’s plot

SectorSlotChallenge

Challenge used for a particular sector for particular slot

SegmentCommitment

Segment commitment contained within segment header.

SegmentIndex

Segment index type.

Solution

Farmer solution for slot challenge.

U256

256-bit unsigned integer

Enums

ArchivedBlockProgress

Progress of an archived block.

NonEmptyVecErr

Error codes for NonEmptyVec.

SegmentHeader

Segment header for a specific segment.

Constants

BLAKE3_HASH_SIZE

Size of BLAKE3 hash output (in bytes).

PUBLIC_KEY_LENGTH

Length of public key in bytes.

RANDOMNESS_LENGTH

Byte length of a randomness type.

REWARD_SIGNATURE_LENGTH

Length of signature in bytes

REWARD_SIGNING_CONTEXT

Signing context used for creating reward signatures by farmers.

Functions

bidirectional_distance

Bidirectional distance metric implemented on top of subtraction

sectors_to_solution_range

Computes the following:

solution_range_to_sectors

Computes the following:

Type Aliases

Blake3Hash

BLAKE3 hash output

BlockHash

Block hash in Subspace network.

BlockNumber

Block number in Subspace network.

BlockWeight

BlockWeight type for fork choice rules.

SectorIndex

Sector index in consensus

SlotNumber

Slot number in Subspace network.

SolutionRange

Type of solution range.