Crate subspace_core_primitives

Core primitives for Subspace Network.

Modules

crypto

Various cryptographic utilities used across Subspace Network.

objects

Data structures related to objects (useful data) stored on Subspace Network.

sector_codec

PoR codec that uses polynomial-based systematic erasure coding

Structs

ArchivedHistorySegment

Archived history segment after archiving is applied.

ChunkSignature

VRF signature output and proof as produced by schnorrkel crate.

FlatPieces

Flat representation of multiple pieces concatenated for higher efficient for processing.

LastArchivedBlock

Last archived block

LegacySectorId

Data structure representing sector ID in farmer’s plot

Piece

A piece of archival history in Subspace Network.

PieceArray

A piece of archival history in Subspace Network.

PieceIndex

Piece index in consensus

PieceIndexHash

Hash of PieceIndex

PosProof

Proof of space proof bytes.

PosQualityBytes

Proof of space quality.

PosSeed

Proof of space seed.

PublicKey

A Ristretto Schnorr public key as bytes produced by schnorrkel crate.

RawRecord

Raw record contained within recorded history segment before archiving is applied.

Record

Record contained within a piece.

RecordCommitment

Record commitment contained within a piece.

RecordWitness

Record witness contained within a piece.

RecordedHistorySegment

Recorded history segment before archiving is applied.

RewardSignature

A Ristretto Schnorr signature as bytes produced by schnorrkel crate.

SegmentIndex

Segment index type.

Solution

Farmer solution for slot challenge.

U256

256-bit unsigned integer

Enums

ArchivedBlockProgress

Progress of an archived block.

SegmentHeader

Segment header for a specific segment.

Constants

BLAKE2B_256_HASH_SIZE

Size of BLAKE2b-256 hash output (in bytes).

PIECES_IN_SECTOR

How many pieces we have in a sector

PLOT_SECTOR_SIZE

Size of one plotted sector.

PUBLIC_KEY_LENGTH

Length of public key in bytes.

RANDOMNESS_CONTEXT

Randomness context

RANDOMNESS_LENGTH

Byte length of a randomness type.

REWARD_SIGNATURE_LENGTH

Length of signature in bytes

Functions

bidirectional_distance

Bidirectional distance metric implemented on top of subtraction

Type Definitions

Blake2b256Hash

BLAKE2b-256 hash output

BlockNumber

Block number in Subspace network.

BlockWeight

BlockWeight type for fork choice rules.

Randomness

Type of randomness.

SectorIndex

Sector index in consensus

SegmentCommitment

Segment commitment type.

SlotNumber

Slot number in Subspace network.

SolutionRange

Type of solution range.